+++ packages/functions/android/src/main/java/io/invertase/firebase/functions/UniversalFirebaseFunctionsModule.java
@@ -1,0 +1,62 @@
+package io.invertase.firebase.functions;
+
+/*
+ * Copyright (c) 2016-present Invertase Limited & Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this library except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+import android.content.Context;
+
+import com.google.android.gms.tasks.Task;
+import com.google.android.gms.tasks.Tasks;
+import com.google.firebase.FirebaseApp;
+import com.google.firebase.functions.FirebaseFunctions;
+
+import java.util.concurrent.Executor;
+
+@SuppressWarnings("ALL")
+public class UniversalFirebaseFunctionsModule {
+  public static final String DATA_KEY = "data";
+  public static final String CODE_KEY = "code";
+  public static final String MSG_KEY = "message";
+  public static final String DETAILS_KEY = "details";
+
+  private final Context context;
+
+  UniversalFirebaseFunctionsModule(Context context) {
+    this.context = context;
+  }
+
+  Task<Object> httpsCallable(
+    String appName,
+    String region,
+    String origin,
+    String name,
+    Object data,
+    Executor executor
+  ) {
+    return Tasks.call(executor, () -> {
+      FirebaseApp firebaseApp = FirebaseApp.getInstance(appName);
+      FirebaseFunctions functionsInstance = FirebaseFunctions.getInstance(firebaseApp, region);
+
+      if (origin != null) {
+        functionsInstance.useFunctionsEmulator(origin);
+      }
+
+      return Tasks.await(functionsInstance.getHttpsCallable(name).call(data)).getData();
+    });
+  }
+
+}
